Parallel Programming working
There is a slow function taking a frequency and calculating the port_data_word before setting it via the Parallel Port. The Fast Function takes the port_data_word directly and sets it via the ParallelPort. This takes only approximately 300ns. We have to see how fast this can get when being triggered.