Test rising edge interrupt for triggering
Takes 336ns up to 424ns and therefore no significant difference in startng delay to while-loop.But the while-loop at the end is not needed and the Interrupt works down to 50ns pulses. Furthermore, the edge Interrupt is independent (at least for two pulses which I can test at home) of the output of the while-loop. This means that two successive parallelPortSettings are 840ns appart (except for the case when the if loop is entered, then it is 1050ns).