Implement Rising Edge Interrupt in Main Code
This makes the delay between Trigger und function setting take 640ns up to 1,3us which is more unstable compared to up to 640ns before. This is by far slower. A reason could be that we do not wait in the while loop directly before until the trigger starts but the while loop is quite big (while dataTransmissionFinished...) with function calls in between. -> reduce code and function calls.
However, two trigger pulses can now be as close as 800ns and two succesive frequency settings are max. 1,35us appart. -> 750kHz update rate could be possible.