Restucture code
Go back to the AD9910 lib inline function which is (as inline function) as fast as the test version with code directly written.
Only speed improvement left is when getting the array value.
Note the currently the while (dataTransmissionFinished)-loop never ends. When the if clause "(arrWriteIndex > arrReadIndex-1)" changing some bools and setting arrWriteIndex =0 is called, the frequency update takes roughly 800ns instead of 700ns. But this is now problem for an experimental cycle as this happens only after the end of the shot.