DMA test
This is how memory-to-memory transfer can work, but the DMA Controller cannot, as I understand, access the PIO registers. Therefore it is not suited for our purposes. DMA can directly write to peripheral devices such as SPI, PWM and in special cases also the DAC but as I get it from the Datasheet page 6 picture, not to PIO.